字符转数字以 _r开头的函数接受double参数

+++ streamingpro-spark-2.0/src/main/java/org/apache/spark/ml/help/HSQLStringIndex.scala
@@ -40,8 +40,8 @@ object HSQLStringIndex {
       }
     }
 
-    sparkSession.udf.register(name + "_r", (index: Int) => {
-      model.value.labels(index)
+    sparkSession.udf.register(name + "_r", (index: Double) => {
+      model.value.labels(index.toInt)
     })
 
     UserDefinedFunction(f, DoubleType, Some(Seq(StringType)))
